% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizbuzz.R
\name{vizbuzz_compare}
\alias{vizbuzz_compare}
\title{Vizbuzz Compare}
\usage{
vizbuzz_compare(path_to_original_image, path_to_replicate_image, fuzz = 10)
}
\arguments{
\item{path_to_original_image}{original image (url, local file path or whatever)}

\item{path_to_replicate_image}{replicated image (url, local file path or whatever)}

\item{fuzz}{acceptable colour distance between 0-100}
}
\description{
Vizbuzz Compare
}
\section{fuzz = relative color distance (value between 0 and 100) to be considered}{
similar in the filling algorithm
see https://imagemagick.org/script/command-line-options.php#fuzz
Change this, re-Knit the file and see what it does in the comparison image
}

